refactor: extract bootstrap mod
