+++ src/bootstrap/jobs/mod.rs
@@ -1,10 +1,14 @@
+pub mod http_tracker;
+pub mod torrent_cleanup;
+pub mod tracker_apis;
+pub mod udp_tracker;
+
 use std::sync::Arc;
 
 use log::warn;
 use tokio::task::JoinHandle;
 use torrust_tracker_configuration::Configuration;
 
-use crate::jobs::{http_tracker, torrent_cleanup, tracker_apis, udp_tracker};
 use crate::servers::http::Version;
 use crate::tracker;
