--- src/jobs/mod.rs
@@ -1,4 +1,0 @@
-pub mod http_tracker;
-pub mod torrent_cleanup;
-pub mod tracker_apis;
-pub mod udp_tracker;
